Add files via upload
This is a to-do list project. You can enter a task and click to add button, your task will be added at the bottom. Task have two options. One to click on done with the symbol of tick mark. This will strike out the task but it will still be there. Other button is to delete the task. Copyright of the picture belongs to me.